+++ src/components/Home/NomadFavouriteDestination/NomadFavouriteDestination.jsx
@@ -1,16 +1,17 @@
 import React, { useEffect, useState } from 'react';
-import destinationsData from "./PopularDestination.json";
+import destinationsData from './NomadFavouriteDestinations.json'
 
-const PopularDestination = () => {
+const NomadFavouriteDestination = () => {
     const [destinations, setDestinations] = useState([]);
 
     useEffect(() => {
         setDestinations(destinationsData);
     }, []);
+
     return (
-        <section className="px-5 md:px-10 lg:px-40 py-10">
+        <section className="px-5 md:px-10 lg:px-20 py-10">
             <h2 className="text-2xl md:text-3xl font-bold mb-8">
-                Popular Destinations
+                Nomads’ Favourite Destinations
             </h2>
 
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
@@ -33,4 +34,4 @@ const PopularDestination = () => {
     );
 };
 
-export default PopularDestination;
+export default NomadFavouriteDestination;
