+++ src/components/Home/PopularDestination/PopularDestination.jsx
@@ -1,17 +1,16 @@
 import React, { useEffect, useState } from 'react';
-import destinationsData from './NomadFavouriteDestinations.json'
+import destinationsData from "./PopularDestination.json";
 
-const NomadFavouriteDestination = () => {
+const PopularDestination = () => {
     const [destinations, setDestinations] = useState([]);
 
     useEffect(() => {
         setDestinations(destinationsData);
     }, []);
-
     return (
-        <section className="px-5 md:px-10 lg:px-40 py-10">
+        <section className="px-5 md:px-10 lg:px-20 py-10">
             <h2 className="text-2xl md:text-3xl font-bold mb-8">
-                Nomads’ Favourite Destinations
+                Popular Destinations
             </h2>
 
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
@@ -34,4 +33,4 @@ const NomadFavouriteDestination = () => {
     );
 };
 
-export default NomadFavouriteDestination;
+export default PopularDestination;
